+++ 103 + 309/Project/Source/main.c
@@ -1,6 +1,7 @@
 #include "main.h"
 #include "Flash64KAppTest.h"
 #include "FactoryAging.h"
+#include "app_lowpower.h"
 
 UINT8 SeriesNum = 10;
 
@@ -133,7 +134,8 @@ int main(void)
 
 		App_Sci();
 		App_AnlogCal();
-		App_LowPowerProcess();
+		// App_LowPowerProcess();
+		LP_Task();
 		App_Can();
 		// App_SleepDeal(); // 关闭这个功能的话，在InitVar()中System_OnOFF_Func相关置零，或者直接屏蔽
 		// App_SOC();